Separate Execution from its Construction
The original impl starts processing once a sink operator is used.
The drawback is that processing cannot be suspended and it fails to support processing of multiple branches.
Now processing logics will be wrapped into execution objects before execution.
The purpose is to make it possible to create different objects for execution in different context.

An execution object contains
+ function `process`: to accept a single input;
+ function `end`: to end the processing.
+ member `control`: to stop processing in the middle.